Restructure + Ack suggestions
